+++ registo.c
@@ -25,7 +25,7 @@ void guarda_custo_no_registo(Registo* registo, int custo_centimos) {
 }
 
 void imprime_saida(Registo* registo) {
-    float custo_euros = registo->custo_centimos / 100;
+    float custo_euros = registo->custo_centimos / (float)100;
     printf("%s %02d-%02d-%02d %02d:%02d %02d-%02d-%02d %02d:%02d %.2f\n",
     registo->carro->matricula, registo->entrada->dia, registo->entrada->mes,
     registo->entrada->ano, registo->entrada->hora, registo->entrada->minutos,
